[model] feat: add wan2.2
