+++ configs/model_configs/wan/wan2_2_i2v_a14b.json
