+++ veomni/models/auto.py
@@ -52,7 +52,9 @@ def build_foundation_model(
     config_path: str,
     weights_path: Optional[str] = None,
     torch_dtype: Literal["float16", "bfloat16", "float32"] = "bfloat16",
-    attn_implementation: Optional[Literal["eager", "sdpa", "flash_attention_2", "flash_attention_3"]] = "flash_attention_2",
+    attn_implementation: Optional[
+        Literal["eager", "sdpa", "flash_attention_2", "flash_attention_3"]
+    ] = "flash_attention_2",
     moe_implementation: Optional[Literal["eager", "fused"]] = None,
     init_device: Literal["cpu", "cuda", "meta"] = "cuda",
     config_kwargs: Optional[Dict[str, Any]] = None,
